Swap Home and Menu
Swapped names for HomeScreen.js and MenuScreen.js
