+++ App.js
@@ -1,8 +1,8 @@
 import { createAppContainer } from 'react-navigation';
 import { createStackNavigator } from 'react-navigation-stack';
 import HomeScreen from "./src/screens/HomeScreen";
-import QuizScreen from "./src/screens/QuizScreen";
 import MenuScreen from './src/screens/MenuScreen';
+import QuizScreen from "./src/screens/QuizScreen";
 import ComicScreen from './src/screens/ComicScreen';
 
 const navigator = createStackNavigator(
@@ -13,7 +13,7 @@ const navigator = createStackNavigator(
     Comic: ComicScreen,
   },
   {
-    initialRouteName: "Menu",
+    initialRouteName: "Home",
     defaultNavigationOptions: {
       title: "App"
     }
